DROID-3013 objects list refactoring
